Adds option for a custom new tab file
Adding the option for a custom new tab. The user can "upload" (rather: load) a file, which is then saved into  browser.storage.local. That way, it can even load external resources, if they are hosted on a website, not via file://.
I added the German and English messages, I do not speak the other two.